requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.25.0

# Database clients
supabase>=2.3.0
neo4j>=5.14.0
redis>=5.0.0

# Google AI
google-generativeai>=0.3.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
